requirements.txt
# 复旦EE招生素材生成器 - 依赖包
# Python 3.10+

# Gemini API
google-genai>=1.68.0

# Claude API (高级优化)
anthropic>=0.86.0

# OpenAI API (GPT精修)
openai>=1.0.0

# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# PDF 处理
pymupdf>=1.24.0

# 长图导出
playwright>=1.40.0

# 其他工具
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
